BDD : changement note en float (int de base)

+++ SEE/sql/Tables.sql
@@ -13,7 +13,7 @@ CREATE TABLE TVolunteer(
     nom VARCHAR(100) NOT NULL,
     age INT CHECK (age>17),
     dpt INT NOT NULL,
-    note INT DEFAULT NULL,
+    note FLOAT DEFAULT NULL,
     nb_avis INT DEFAULT 0,
     psw VARCHAR(100) NOT NULL
 );
